+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 + Fixed Flatten to sneak to place on blocks that need to be sneaken on
 + Fixed other entities render rotations from being 0'd out when using spoof rotations
 + Add more options to Animations module
++ Add Color options to Interface
 ```
 
 ---